Always add the sort argument when it exists.
The query for max_day uses find_one and with a sort parameter
to get the latest; however using find_one was causing the sort
parameter to be ignored.

Instead, always use the sort parameter for both find and find_one
queries.

This was discovered when preparing the findb dump for use with the S3
securities folder, which uses the max_day as part of the url to S3. The
max_day found was the first max_day in the system, not the last max_day;
because the sort was omitted, causing an error because the S3 data only
goes back to 2014-07-08.